+++ modules/_faux/CustomCraftVariable.php
@@ -4,6 +4,7 @@ namespace modules\_faux;
 
 use craft\web\twig\variables\CraftVariable;
 use modules\main\services\ProjectService;
+use nystudio107\vite\variables\ViteVariable;
 use spacecatninja\imagerx\variables\ImagerVariable;
 use Spatie\SchemaOrg\Schema;
 
@@ -11,6 +12,7 @@ use Spatie\SchemaOrg\Schema;
  * @mixin Schema $schema
  * @mixin ImagerVariable $imager
  * @mixin ProjectService $project
+ * @mixin ViteVariable $vite
  */
 
 class CustomCraftVariable extends CraftVariable
@@ -28,5 +30,10 @@ class CustomCraftVariable extends CraftVariable
     public function project(): ProjectService
     {
         return new ProjectService();
+    }
+
+    public function vite(): ViteVariable
+    {
+        return new ViteVariable();
     }
 }